Adding functions for AirlightEstimation and BoundCon
AirlightEstimation(self, HazeImg): This method estimates the atmospheric light in a hazy image. It takes an input image HazeImg as a parameter. The method first checks the shape of the image to determine if it is a color image or a grayscale image.

BoundCon(self, HazeImg): This method calculates the transmission map of the hazy image based on the estimated atmospheric light and predefined constants (self.C0 and self.C1). It takes the hazy image HazeImg as a parameter.